Add ability to choose task for ex2
