+++ lab2/include/gamma_correction.hpp
@@ -1,8 +1,9 @@
+#pragma once
 #include <opencv2/opencv.hpp>
 
 namespace tud::cvlabs
 {
-    void gammaCurveMain(cv::Mat &image);
-
     auto applyGammaCorrection(const cv::Mat &image, double gamma);
+
+    void gammaCurveMain(cv::Mat &image);
 }
