+++ lab2/include/hist_equalization.hpp
@@ -1,3 +1,4 @@
+#pragma once
 #include <opencv2/opencv.hpp>
 
 namespace tud::cvlabs
@@ -5,4 +6,6 @@ namespace tud::cvlabs
     cv::Mat applyEqualization(const cv::Mat &image, cv::Mat &H);
 
     cv::Mat calculateHistogram(const cv::Mat &image);
+
+    void histEqualMain(const cv::Mat &image);
 }
