+++ lab2/include/sobel_filter.hpp
@@ -1,7 +1,10 @@
+#pragma once
 #include <opencv2/opencv.hpp>
 
 namespace tud::cvlabs
 {
     cv::Mat applySobel(const cv::Mat &image, uint ddepth = CV_8U,
                        cv::BorderTypes borderType = cv::BorderTypes::BORDER_REFLECT);
+
+    void sobelMain(const cv::Mat &image);
 }
