+++ lab2/src/gamma_correction.cpp
@@ -1,4 +1,5 @@
 #include <iostream>
+#include <tuple>
 
 #include <opencv2/highgui.hpp>
