+++ lab2/src/hist_equalization.cpp
@@ -1,13 +1,16 @@
+#include <cmath>
+#include <vector>
 #include <numeric>
+#include <algorithm>
 
 #include "hist_equalization.hpp"
 
 namespace tud::cvlabs
 {
-    cv::Mat applyEqualization(const cv::Mat &image, cv::Mat &H)
+    cv::Mat applyEqualization(const cv::Mat &image, cv::Mat &hist)
     {
         std::vector<double> cdf;
-        std::inclusive_scan(H.begin<double>(), H.end<double>(),
+        std::inclusive_scan(hist.begin<double>(), hist.end<double>(),
                             std::back_inserter(cdf), std::plus<double>{});
 
         std::vector<uchar> lut;
@@ -57,5 +60,46 @@ namespace tud::cvlabs
         cv::merge(resChannels, result);
 
         return result;
+    }
+
+    void showHist(const cv::Mat &hist)
+    {
+        std::vector<uchar> y;
+
+        double max = *std::max_element(hist.begin<double>(), hist.end<double>());
+        double norm = (255 - 20) / max;
+
+        std::transform(hist.begin<double>(), hist.end<double>(), std::back_inserter(y),
+                       [norm](double h)
+                       {
+                           return cv::saturate_cast<uchar>(h * norm);
+                       });
+
+        cv::Mat result = cv::Mat::zeros(256, 256, CV_8U);
+        for (int i = 0; i < 256; ++i)
+        {
+            cv::rectangle(result, cv::Rect(i, 255 - y[i], 1, y[i]), 255);
+        }
+
+        imshow("Histogram", result);
+    }
+
+    void histEqualMain(const cv::Mat &image)
+    {
+        cv::Mat grayImg;
+        cv::cvtColor(image, grayImg, cv::COLOR_BGR2GRAY);
+
+        auto hist = calculateHistogram(grayImg);
+        auto result = applyEqualization(grayImg, hist);
+
+        showHist(hist);
+        imshow("Our Hist Equalization", result);
+        imshow("Original Image", grayImg);
+
+        cv::Mat reference;
+        cv::equalizeHist(grayImg, reference);
+        imshow("Reference Hist Equalization", reference);
+
+        cv::waitKey();
     }
 }
